+++ package.json
@@ -53,7 +53,7 @@
   ],
   "dependencies": {
     "@types/googlemaps": "^3.39.6",
-    "algoliasearch-helper": "^3.2.2",
+    "algoliasearch-helper": "^3.3.2",
     "classnames": "^2.2.5",
     "events": "^1.1.0",
     "hogan.js": "^3.0.2",
@@ -62,7 +62,7 @@
     "qs": "^6.5.1"
   },
   "devDependencies": {
-    "@algolia/client-search": "4.3.1",
+    "@algolia/client-search": "4.8.2",
     "@babel/cli": "7.8.3",
     "@babel/core": "7.8.3",
     "@babel/plugin-proposal-class-properties": "7.8.3",
@@ -92,7 +92,7 @@
     "@wdio/selenium-standalone-service": "5.16.5",
     "@wdio/spec-reporter": "5.16.5",
     "@wdio/static-server-service": "5.16.5",
-    "algoliasearch": "4.3.1",
+    "algoliasearch": "4.8.2",
     "algoliasearch-v3": "npm:algoliasearch@3",
     "babel-eslint": "10.0.3",
     "babel-jest": "24.9.0",
